Implement BlockRegistrar for audio player and refactor BlockServiceProvider
