+++ plugins/bifrost-player/inc/Block/BlockServiceProvider.php
@@ -16,10 +16,8 @@ namespace Bifrost\Player\Block;
 use Bifrost\Framework\Contracts\Bootable;
 use Bifrost\Framework\Core\ServiceProvider;
 
-use const Bifrost\Player\PLUGIN_DIR;
-
 /**
- * Registers interactive blocks and the playlist track render filter.
+ * Registers block-related services and boots them.
  */
 final class BlockServiceProvider extends ServiceProvider implements Bootable
 {
@@ -28,27 +26,9 @@ final class BlockServiceProvider extends ServiceProvider implements Bootable
 	 */
 	public function boot(): void
 	{
-		add_action('init', $this->registerBlocks(...));
-		add_action('wp_footer', $this->renderPlayer(...));
-
+		$this->container->get(BlockRegistrar::class)->boot();
 		$this->container->get(RenderPlayButton::class)->boot();
 		$this->container->get(RenderPlaylist::class)->boot();
 		$this->container->get(RenderPlaylistTrack::class)->boot();
-	}
-
-	/**
-	 * Registers the plugin's interactive blocks.
-	 */
-	private function registerBlocks(): void
-	{
-		register_block_type(PLUGIN_DIR . '/build/blocks/audio-player');
-	}
-
-	/**
-	 * Renders the persistent audio player in the footer.
-	 */
-	private function renderPlayer(): void
-	{
-		echo do_blocks('<!-- wp:bifrost-player/audio-player /-->');
 	}
 }
